+++ mobile/android/android-components/components/browser/menu/src/main/java/mozilla/components/browser/menu/BrowserMenuItem.kt
@@ -7,6 +7,7 @@ package mozilla.components.browser.menu
 import android.content.Context
 import android.view.View
 import mozilla.components.browser.menu.view.ExpandableLayout
+import mozilla.components.browser.menu.view.StickyItemsLinearLayoutManager
 import mozilla.components.concept.menu.candidate.MenuCandidate
 
 /**
@@ -32,6 +33,13 @@ interface BrowserMenuItem {
      * @see [ExpandableLayout]
      */
     val isCollapsingMenuLimit: Boolean get() = false
+
+    /**
+     * Whether this menu item should not be scrollable off-screen.
+     *
+     * @see [StickyItemsLinearLayoutManager]
+     */
+    val isSticky: Boolean get() = false
 
     /**
      * Returns the layout resource ID of the layout to be inflated for showing a menu item of this