+++ mobile/android/android-components/components/browser/menu/src/main/java/mozilla/components/browser/menu/item/BrowserMenuHighlightableSwitch.kt
@@ -25,6 +25,8 @@ import mozilla.components.concept.menu.candidate.LowPriorityHighlightEffect
  *
  * @param label The visible label of this menu item.
  * @param isCollapsingMenuLimit Whether this menu item can serve as the limit of a collapsing menu.
+ * @param isSticky whether this item menu should not be scrolled offscreen (downwards or upwards
+ * depending on the menu position).
  * @param initialState The initial value the checkbox should have.
  * @param listener Callback to be invoked when this menu item is checked.
  */
@@ -35,11 +37,12 @@ class BrowserMenuHighlightableSwitch(
     @ColorRes private val iconTintColorResource: Int = NO_ID,
     @ColorRes private val textColorResource: Int = NO_ID,
     override val isCollapsingMenuLimit: Boolean = false,
+    override val isSticky: Boolean = false,
     override val highlight: BrowserMenuHighlight.LowPriority,
     override val isHighlighted: () -> Boolean = { true },
     initialState: () -> Boolean = { false },
     listener: (Boolean) -> Unit
-) : BrowserMenuCompoundButton(label, isCollapsingMenuLimit, initialState, listener), HighlightableMenuItem {
+) : BrowserMenuCompoundButton(label, isCollapsingMenuLimit, isSticky, initialState, listener), HighlightableMenuItem {
 
     private var wasHighlighted = false