+++ mobile/android/android-components/components/browser/menu/src/main/java/mozilla/components/browser/menu/item/BrowserMenuImageSwitch.kt
@@ -21,18 +21,22 @@ import java.lang.reflect.Modifier
  * @param imageResource ID of a drawable resource to be shown as icon.
  * @param label The visible label of this menu item.
  * @param isCollapsingMenuLimit Whether this menu item can serve as the limit of a collapsing menu.
+ * @param isSticky whether this item menu should not be scrolled offscreen (downwards or upwards
+ * depending on the menu position).
  * @param initialState The initial value the checkbox should have.
  * @param listener Callback to be invoked when this menu item is checked.
  */
+@Suppress("LongParameterList")
 class BrowserMenuImageSwitch(
     @VisibleForTesting(otherwise = Modifier.PRIVATE)
     @DrawableRes
     val imageResource: Int,
     label: String,
     override val isCollapsingMenuLimit: Boolean = false,
+    override val isSticky: Boolean = false,
     initialState: () -> Boolean = { false },
     listener: (Boolean) -> Unit
-) : BrowserMenuCompoundButton(label, isCollapsingMenuLimit, initialState, listener) {
+) : BrowserMenuCompoundButton(label, isCollapsingMenuLimit, isSticky, initialState, listener) {
     override fun getLayoutResource(): Int = R.layout.mozac_browser_menu_item_image_switch
 
     override fun bind(menu: BrowserMenu, view: View) {