+++ mobile/android/android-components/components/browser/menu/src/main/java/mozilla/components/browser/menu/item/BrowserMenuImageTextCheckboxButton.kt
@@ -24,6 +24,9 @@ import mozilla.components.support.ktx.android.util.dpToPx
  * @param labelListener Callback to be invoked when this menu item is clicked.
  * @param primaryStateIconResource ID of a drawable resource for checkbox drawable in primary state.
  * @param secondaryStateIconResource ID of a drawable resource for checkbox drawable in secondary state.
+ * @param isCollapsingMenuLimit Whether this menu item can serve as the limit of a collapsing menu.
+ * @param isSticky whether this item menu should not be scrolled offscreen (downwards or upwards
+ * depending on the menu position).
  * @param iconTintColorResource Optional ID of color resource to tint the checkbox drawable.
  * @param primaryLabel The visible label of the checkbox in primary state.
  * @param secondaryLabel The visible label of this menu item in secondary state.
@@ -43,10 +46,11 @@ class BrowserMenuImageTextCheckboxButton(
     private val primaryLabel: String,
     private val secondaryLabel: String,
     override val isCollapsingMenuLimit: Boolean = false,
+    override val isSticky: Boolean = false,
     val isInPrimaryState: () -> Boolean = { true },
     private val onCheckedChangedListener: (Boolean) -> Unit
 ) : BrowserMenuImageText(
-    label, imageResource, iconTintColorResource, textColorResource, isCollapsingMenuLimit, labelListener
+    label, imageResource, iconTintColorResource, textColorResource, isCollapsingMenuLimit, isSticky, labelListener
 ) {
     override var visible: () -> Boolean = { true }
     override fun getLayoutResource(): Int = R.layout.mozac_browser_menu_item_image_text_checkbox_button